fix: incorrect cast

+++ imu/imu.c
@@ -417,9 +417,9 @@ if ( imu_status != IMU_OK )
     gyro_y_raw = ( (uint16_t) regGyro[0] ) << 8 | regGyro[1];
     gyro_z_raw = ( (uint16_t) regGyro[0] ) << 8 | regGyro[1];
 #elif defined( A0002_REV2 )
-    gyro_x_raw = ( (int16_t) regGyro[1] ) << 8 | regGyro[0];
-    gyro_y_raw = ( (int16_t) regGyro[3] ) << 8 | regGyro[2];
-    gyro_z_raw = ( (int16_t) regGyro[5] ) << 8 | regGyro[4];
+    gyro_x_raw = (int16_t) ( (uint16_t) regGyro[1] << 8 | regGyro[0] );
+    gyro_y_raw = (int16_t) ( (uint16_t) regGyro[3] << 8 | regGyro[2] );
+    gyro_z_raw = (int16_t) ( (uint16_t) regGyro[5] << 8 | regGyro[4] );
 #endif
 
 /* Export Sensor Readouts */
@@ -477,12 +477,12 @@ if ( imu_status != IMU_OK )
 	}
 
 /* Combine high byte and low byte to 16 bit data  */
-accel_x_raw = ( (int16_t) regRaw[1] ) << 8 | regRaw[0];
-accel_y_raw = ( (int16_t) regRaw[3] ) << 8 | regRaw[2];
-accel_z_raw = ( (int16_t) regRaw[5] ) << 8 | regRaw[4];
-gyro_x_raw = ( (int16_t) regRaw[7] ) << 8 | regRaw[6];
-gyro_y_raw = ( (int16_t) regRaw[9] ) << 8 | regRaw[8];
-gyro_z_raw = ( (int16_t) regRaw[11] ) << 8 | regRaw[10];
+accel_x_raw = (int16_t) ( (uint16_t) regRaw[1] << 8 | regRaw[0] );
+accel_y_raw = (int16_t) ( (uint16_t) regRaw[3] << 8 | regRaw[2] );
+accel_z_raw = (int16_t) ( (uint16_t) regRaw[5] << 8 | regRaw[4] );
+gyro_x_raw  = (int16_t) ( (uint16_t) regRaw[7] << 8 | regRaw[6] );
+gyro_y_raw  = (int16_t) ( (uint16_t) regRaw[9] << 8 | regRaw[8] );
+gyro_z_raw  = (int16_t) ( (uint16_t) regRaw[11] << 8 | regRaw[10] );
 
 /* Export Sensor Readouts */
 pIMU->accel_x = accel_x_raw;
@@ -1137,12 +1137,12 @@ if( !imu_data_ready && !mag_data_ready )
     Handle IT signal
     ------------------------------------------------------------------------------*/
     /* Combine high byte and low byte to 16 bit data  */
-    imu_raw_processed.accel_x = ( (int16_t) imu_raw_buffer[1] ) << 8 | imu_raw_buffer[0];
-    imu_raw_processed.accel_y = ( (int16_t) imu_raw_buffer[3] ) << 8 | imu_raw_buffer[2];
-    imu_raw_processed.accel_z = ( (int16_t) imu_raw_buffer[5] ) << 8 | imu_raw_buffer[4];
-    imu_raw_processed.gyro_x = ( (int16_t) imu_raw_buffer[7] ) << 8 | imu_raw_buffer[6];
-    imu_raw_processed.gyro_y = ( (int16_t) imu_raw_buffer[9] ) << 8 | imu_raw_buffer[8];
-    imu_raw_processed.gyro_z = ( (int16_t) imu_raw_buffer[11] ) << 8 | imu_raw_buffer[10];
+    imu_raw_processed.accel_x = (int16_t) ( (uint16_t) imu_raw_buffer[1] << 8 | imu_raw_buffer[0] );
+    imu_raw_processed.accel_y = (int16_t) ( (uint16_t) imu_raw_buffer[3] << 8 | imu_raw_buffer[2] );
+    imu_raw_processed.accel_z = (int16_t) ( (uint16_t) imu_raw_buffer[5] << 8 | imu_raw_buffer[4] );
+    imu_raw_processed.gyro_x  = (int16_t) ( (uint16_t) imu_raw_buffer[7] << 8 | imu_raw_buffer[6] );
+    imu_raw_processed.gyro_y  = (int16_t) ( (uint16_t) imu_raw_buffer[9] << 8 | imu_raw_buffer[8] );
+    imu_raw_processed.gyro_z  = (int16_t) ( (uint16_t) imu_raw_buffer[11] << 8 | imu_raw_buffer[10] );
 
     imu_data_ready = true;
